pyproject.toml
[project]
name = "risk-tier-tool"
version = "0.2.0"
description = "A tool to assess and visualize risk tiers based on user input."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.5",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pydantic>=2.12.4",
    "streamlit==1.52.0",
    "streamlit-antd-components>=0.3.2",
    "streamlit-code-editor>=0.1.22",
    "streamlit-flow-component>=1.6.1",
    "streamlit-sortables>=0.3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.11.13",
]

[tool.ruff.format]
line-ending = "cr-lf"
